+++ CHANGELOG.md
@@ -14,6 +14,7 @@ plus de détails sur http://semver.org/lang/fr/
 * block `focus` :
 	* correction de la taille des composants `embed`
 	* centrage des composants `image` placés en haut
+* block `mosaic` ajouté
 * composant `buttons` : champ passé en optionnel dans ACF
 * composant `image` : ajout des crédits
 * composant `links` :